Add PaymentCard functionality & implement Banorte encryption feature
This commit focuses on enhancing the PaymentCard component by adding properties that aid in payment processing. A new function 'empezarBanorte' is added to facilitate this process. Additionally, a 'banorte.ts' file is created to support encryption operations required for payment. Changes also include the addition of a 'randomByte' import from 'crypto', a function to generate a control number, and updates to the '.gitignore' and 'package.json' files.