feat: Update num2words.ts
Updated the num2words.ts file to handle a special case where '1' is converted to 'un peso' instead of 'uno peso'. Also added handling for pluralization of 'peso' and 'centavo' based on the value of the integerPart and cents variables.

This resolves the issue of incorrect output for the amount in words.